requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
cryptography==2.9.2
Django==3.2.9
django-cors-headers==3.1.1
django-environ==0.8.1
djangorestframework==3.12.4
drf-jwt==1.13.4
idna==2.8
pycparser==2.21
PyJWT==1.7.1
pytz==2021.3
requests==2.22.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.25.11
